+++ src/scr/libs/StuntCarRacerSharpLib/StuntCarRacerMain.cs
@@ -105,7 +105,7 @@ public sealed class StuntCarRacerMain : IGame, IGameApp
         Screens = new(Keyboard);
         Screens.Add(GameMode.TrackMenu, new TrackMenuScreen(Race, Keyboard, Screens, Graphics, Layout, Palette));
         Screens.Add(GameMode.TrackPreview, new TrackPreviewScreen(Race, Keyboard, Screens, Graphics, Layout, Palette));
-        Screens.Add(GameMode.GameInProgress, new RaceScreen(Race, Keyboard, Screens));
+        Screens.Add(GameMode.GameInProgress, new RaceScreen(Race, Keyboard, Sound, Screens));
         Screens.Add(GameMode.GameOver, new GameOverScreen(Race, Keyboard, Sound, Screens));
         Screens.Set(GameMode.TrackMenu);
     }